Create services with shared provider and network access
