+++ hooks/useHistoricCommits/index.tsx
@@ -1,12 +1,11 @@
 import { useEffect, useState } from 'react';
 import shallow from 'zustand/shallow';
-import { CommitActionEnum, NETWORKS } from '@tracer-protocol/pools-js';
+import { CommitActionEnum } from '@tracer-protocol/pools-js';
 import { CommitActionToQueryFocusMap } from '~/constants/commits';
+import { tracerAPIService } from '~/services/TracerAPIService';
 import { useStore } from '~/store/main';
 import { selectWeb3Info } from '~/store/Web3Slice';
 import { TradeHistory } from '~/types/commits';
-import { V2_SUPPORTED_NETWORKS } from '~/types/networks';
-import { fetchCommitHistory } from '~/utils/tracerAPI';
 import { PAGE_ENTRIES } from '../usePagination';
 
 export const useHistoricCommits = (
@@ -27,17 +26,18 @@ export const useHistoricCommits = (
     useEffect(() => {
         setLoading(true);
         if (account) {
-            fetchCommitHistory({
-                account: account ?? '0',
-                network: (network as V2_SUPPORTED_NETWORKS) ?? NETWORKS.ARBITRUM,
-                type: CommitActionToQueryFocusMap[focus as CommitActionEnum],
-                page,
-                pageSize: PAGE_ENTRIES, // TODO: allow user to choose results per page
-            }).then((r) => {
-                setLoading(false);
-                setTradeHistory(r.results);
-                setTotalRecords(r.totalRecords);
-            });
+            tracerAPIService
+                .fetchCommitHistory({
+                    account: account ?? '0',
+                    type: CommitActionToQueryFocusMap[focus as CommitActionEnum],
+                    page,
+                    pageSize: PAGE_ENTRIES, // TODO: allow user to choose results per page
+                })
+                .then((r) => {
+                    setLoading(false);
+                    setTradeHistory(r.results);
+                    setTotalRecords(r.totalRecords);
+                });
         }
     }, [focus, page, account, network]);
