+++ hooks/useUpkeeps/index.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react';
 import { ethers } from 'ethers';
 import BigNumber from 'bignumber.js';
 import { KnownNetwork, calcSkew, calcTokenPrice } from '@tracer-protocol/pools-js';
+import { last2UpkeepsQuery, subgraphUrlByNetwork } from '~/services/TracerAPIService/subgraph';
 import { V2_SUPPORTED_NETWORKS } from '~/types/networks';
-import { last2UpkeepsQuery, subgraphUrlByNetwork } from '~/utils/tracerAPI/subgraph';
 import { useAllPoolLists } from '../useAllPoolLists';
 
 export type Upkeep = {
